+++ app/strategies/booking_strategies/default/actions/manage/email_invoice.rb
@@ -6,7 +6,7 @@ module BookingStrategies
       module Manage
         class EmailInvoice < BookingStrategy::Action
           def call!(invoices = booking.invoices.unsent)
-            booking.notifications.new(from_template: :payment_due_message, addressed_to: :tenant)&.tap do |notification|
+            booking.notifications.new(from_template: :payment_due, addressed_to: :tenant)&.tap do |notification|
               notification.attachments.attach(invoices.map { |invoice| invoice.pdf.blob })
             end&.deliver! && invoices.each(&:sent!)
           end